+++ axelor-account/src/main/java/com/axelor/apps/account/service/move/MoveService.java
@@ -189,8 +189,6 @@ public class MoveService {
 	 * 		- le compte des dûs est le même que celui de l'avoir : alors on lettre directement
 	 *  	- le compte n'est pas le même : on créée une O.D. de passage sur le bon compte
 	 * @param invoice
-	 * @param company
-	 * @param useExcessPayment
 	 * @return
 	 * @throws AxelorException
 	 */
@@ -202,16 +200,13 @@ public class MoveService {
 
 		Move move = null;
 
-		List<MoveLine> debitMoveLines = Lists.newArrayList();
-		
-		AccountConfig accountConfig = accountConfigService.getAccountConfig(company);
-		
-		if(accountConfig.getAutoReconcileOnInvoice())  {		
-			// Récupération des dûs
-			debitMoveLines.addAll(moveDueService.getInvoiceDue(invoice, true));
-		}
 
-		if(debitMoveLines != null && debitMoveLines.size() != 0)  {
+		AccountConfig accountConfig = accountConfigService.getAccountConfig(company);
+
+		// Récupération des dûs
+		List<MoveLine> debitMoveLines = moveDueService.getInvoiceDue(invoice, accountConfig.getAutoReconcileOnInvoice());
+
+		if(!debitMoveLines.isEmpty())  {
 			MoveLine invoiceCustomerMoveLine = moveToolService.getCustomerMoveLineByLoop(invoice);
 			
 			// Si c'est le même compte sur les trop-perçus et sur la facture, alors on lettre directement