+++ axelor-client-portal/src/main/java/com/axelor/apps/portal/translation/ITranslation.java
@@ -1,0 +1,7 @@
+package com.axelor.apps.portal.translation;
+
+public interface ITranslation {
+	
+	public static final String PORTAL_APP_NAME = /*$$(*/ "value:Portal"; /*)*/
+
+}