+++ axelor-production/src/main/java/com/axelor/apps/production/service/OperationOrderStockMoveService.java
@@ -39,18 +39,19 @@ import com.google.inject.Inject;
 
 public class OperationOrderStockMoveService {
 
+	protected StockMoveService stockMoveService;
+	protected StockMoveLineService stockMoveLineService;
+	protected ProductionConfigService productionConfigService;
+	protected StockLocationRepository stockLocationRepo;
+	
 	@Inject
-	private StockMoveService stockMoveService;
-
-	@Inject
-	private StockMoveLineService stockMoveLineService;
-
-	@Inject
-	private ProductionConfigService productionConfigService;
-
-	@Inject
-	private StockLocationRepository stockLocationRepo;
-
+	public OperationOrderStockMoveService(StockMoveService stockMoveService, StockMoveLineService stockMoveLineService,
+			ProductionConfigService productionConfigService, StockLocationRepository stockLocationRepo)  {
+		this.stockMoveService = stockMoveService;
+		this.stockMoveLineService = stockMoveLineService;
+		this.productionConfigService = productionConfigService;
+		this.stockLocationRepo = stockLocationRepo;
+	}
 
 	public void createToConsumeStockMove(OperationOrder operationOrder) throws AxelorException {
 
@@ -84,7 +85,7 @@ public class OperationOrderStockMoveService {
 	}
 
 
-	private StockMove _createToConsumeStockMove(OperationOrder operationOrder, Company company) throws AxelorException  {
+	protected StockMove _createToConsumeStockMove(OperationOrder operationOrder, Company company) throws AxelorException  {
 
 		StockConfigProductionService stockConfigService = Beans.get(StockConfigProductionService.class);
 		StockConfig stockConfig = stockConfigService.getStockConfig(company);
@@ -109,7 +110,7 @@ public class OperationOrderStockMoveService {
 
 
 
-	private StockMoveLine _createStockMoveLine(ProdProduct prodProduct, StockMove stockMove) throws AxelorException  {
+	protected StockMoveLine _createStockMoveLine(ProdProduct prodProduct, StockMove stockMove) throws AxelorException  {
 
 		return stockMoveLineService.createStockMoveLine(
 				prodProduct.getProduct(),