+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/exception/IExceptionMessage.java
@@ -132,6 +132,11 @@ public interface IExceptionMessage {
 	static final String SUPPLY_CHAIN_CONFIG = /*$$(*/ "You must configure a Supply chain module for the company %s" /*)*/;
 
 	String SUPPLYCHAIN_MISSING_CANCEL_REASON_ON_CHANGING_SALE_ORDER = /*$$(*/ "You must configure a cancel reason on changing sale order in app supplychain." /*)*/;
-
+	
+	/**
+	 * Subscription invoice
+	 */
+	static final String TOTAL_SUBSCRIPTION_INVOICE_GENERATED = /*$$(*/ "Total subscription invoice(s) generated: %s" /*)*/;
+	static final String SUBSCRIPTION_INVOICE_GENERATION_ERROR = /*$$(*/ "Error generating subscription invoice(s): '%s'" /*)*/;
 }